+++ composeApp/src/commonMain/kotlin/com/github/singularity/core/shared/model/PairCheck.kt
@@ -3,7 +3,7 @@ package com.github.singularity.core.shared.model
 import com.github.singularity.core.shared.model.http.PairStatus
 
 data class PairCheck(
-    val requestId: Long,
+    val requestId: Int,
     val node: Node,
     val status: PairStatus = PairStatus.Awaiting,
 )
